PythonProject/requirements.txt
Django==2.2
django-tinymce==2.6.0
mysqlclient==2.0.1
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
